executor changes compiles

+++ crates/core/executor/src/events/instr.rs
@@ -13,15 +13,15 @@ pub struct AluEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The program counter.
-    pub pc: u32,
+    pub pc: u64,
     /// The opcode.
     pub opcode: Opcode,
     /// The first operand value.
-    pub a: u32,
+    pub a: u64,
     /// The second operand value.
-    pub b: u32,
+    pub b: u64,
     /// The third operand value.
-    pub c: u32,
+    pub c: u64,
     /// Whether the first operand is register 0.
     pub op_a_0: bool,
 }
@@ -30,7 +30,7 @@ impl AluEvent {
     /// Create a new [`AluEvent`].
     #[must_use]
     #[allow(clippy::too_many_arguments)]
-    pub fn new(clk: u32, pc: u32, opcode: Opcode, a: u32, b: u32, c: u32, op_a_0: bool) -> Self {
+    pub fn new(clk: u32, pc: u64, opcode: Opcode, a: u64, b: u64, c: u64, op_a_0: bool) -> Self {
         Self { clk, pc, opcode, a, b, c, op_a_0 }
     }
 }
@@ -46,15 +46,15 @@ pub struct MemInstrEvent {
     /// The clk.
     pub clk: u32,
     /// The program counter.
-    pub pc: u32,
+    pub pc: u64,
     /// The opcode.
     pub opcode: Opcode,
     /// The first operand value.
-    pub a: u32,
+    pub a: u64,
     /// The second operand value.
-    pub b: u32,
+    pub b: u64,
     /// The third operand value.
-    pub c: u32,
+    pub c: u64,
     /// Whether the first operand is register 0.
     pub op_a_0: bool,
     /// The memory access record for memory operations.
@@ -68,11 +68,11 @@ impl MemInstrEvent {
     pub fn new(
         shard: u32,
         clk: u32,
-        pc: u32,
+        pc: u64,
         opcode: Opcode,
-        a: u32,
-        b: u32,
-        c: u32,
+        a: u64,
+        b: u64,
+        c: u64,
         op_a_0: bool,
         mem_access: MemoryRecordEnum,
     ) -> Self {
@@ -89,17 +89,17 @@ pub struct BranchEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The program counter.
-    pub pc: u32,
+    pub pc: u64,
     /// The next program counter.
-    pub next_pc: u32,
+    pub next_pc: u64,
     /// The opcode.
     pub opcode: Opcode,
     /// The first operand value.
-    pub a: u32,
+    pub a: u64,
     /// The second operand value.
-    pub b: u32,
+    pub b: u64,
     /// The third operand value.
-    pub c: u32,
+    pub c: u64,
     /// Whether the first operand is register 0.
     pub op_a_0: bool,
 }
@@ -110,12 +110,12 @@ impl BranchEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         clk: u32,
-        pc: u32,
-        next_pc: u32,
+        pc: u64,
+        next_pc: u64,
         opcode: Opcode,
-        a: u32,
-        b: u32,
-        c: u32,
+        a: u64,
+        b: u64,
+        c: u64,
         op_a_0: bool,
     ) -> Self {
         Self { clk, pc, next_pc, opcode, a, b, c, op_a_0 }
@@ -131,17 +131,17 @@ pub struct JumpEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The program counter.
-    pub pc: u32,
+    pub pc: u64,
     /// The next program counter.
-    pub next_pc: u32,
+    pub next_pc: u64,
     /// The opcode.
     pub opcode: Opcode,
     /// The first operand value.
-    pub a: u32,
+    pub a: u64,
     /// The second operand value.
-    pub b: u32,
+    pub b: u64,
     /// The third operand value.
-    pub c: u32,
+    pub c: u64,
     /// Whether the first operand is register 0.
     pub op_a_0: bool,
 }
@@ -152,12 +152,12 @@ impl JumpEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         clk: u32,
-        pc: u32,
-        next_pc: u32,
+        pc: u64,
+        next_pc: u64,
         opcode: Opcode,
-        a: u32,
-        b: u32,
-        c: u32,
+        a: u64,
+        b: u64,
+        c: u64,
         op_a_0: bool,
     ) -> Self {
         Self { clk, pc, next_pc, opcode, a, b, c, op_a_0 }
@@ -172,15 +172,15 @@ pub struct AUIPCEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The program counter.
-    pub pc: u32,
+    pub pc: u64,
     /// The opcode.
     pub opcode: Opcode,
     /// The first operand value.
-    pub a: u32,
+    pub a: u64,
     /// The second operand value.
-    pub b: u32,
+    pub b: u64,
     /// The third operand value.
-    pub c: u32,
+    pub c: u64,
     /// Whether the first operand is register 0.
     pub op_a_0: bool,
 }
@@ -189,7 +189,7 @@ impl AUIPCEvent {
     /// Create a new [`AUIPCEvent`].
     #[must_use]
     #[allow(clippy::too_many_arguments)]
-    pub fn new(clk: u32, pc: u32, opcode: Opcode, a: u32, b: u32, c: u32, op_a_0: bool) -> Self {
+    pub fn new(clk: u32, pc: u64, opcode: Opcode, a: u64, b: u64, c: u64, op_a_0: bool) -> Self {
         Self { clk, pc, opcode, a, b, c, op_a_0 }
     }
 }
