+++ crates/core/executor/src/events/precompiles/fptower.rs
@@ -26,13 +26,13 @@ pub struct FpOpEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The pointer to the x operand.
-    pub x_ptr: u32,
+    pub x_ptr: u64,
     /// The x operand.
-    pub x: Vec<u32>,
+    pub x: Vec<u64>,
     /// The pointer to the y operand.
-    pub y_ptr: u32,
+    pub y_ptr: u64,
     /// The y operand.
-    pub y: Vec<u32>,
+    pub y: Vec<u64>,
     /// The operation to perform.
     pub op: FieldOperation,
     /// The memory records for the x operand.
@@ -55,13 +55,13 @@ pub struct Fp2AddSubEvent {
     /// The operation to perform.
     pub op: FieldOperation,
     /// The pointer to the x operand.
-    pub x_ptr: u32,
+    pub x_ptr: u64,
     /// The x operand.
-    pub x: Vec<u32>,
+    pub x: Vec<u64>,
     /// The pointer to the y operand.
-    pub y_ptr: u32,
+    pub y_ptr: u64,
     /// The y operand.
-    pub y: Vec<u32>,
+    pub y: Vec<u64>,
     /// The memory records for the x operand.
     pub x_memory_records: Vec<MemoryWriteRecord>,
     /// The memory records for the y operand.
@@ -78,13 +78,13 @@ pub struct Fp2MulEvent {
     /// The clock cycle.
     pub clk: u32,
     /// The pointer to the x operand.
-    pub x_ptr: u32,
+    pub x_ptr: u64,
     /// The x operand.
-    pub x: Vec<u32>,
+    pub x: Vec<u64>,
     /// The pointer to the y operand.
-    pub y_ptr: u32,
+    pub y_ptr: u64,
     /// The y operand.
-    pub y: Vec<u32>,
+    pub y: Vec<u64>,
     /// The memory records for the x operand.
     pub x_memory_records: Vec<MemoryWriteRecord>,
     /// The memory records for the y operand.
