+++ crates/core/executor/src/syscalls/deferred.rs
@@ -6,9 +6,9 @@ use super::{SyscallCode, SyscallContext};
 pub fn commit_deferred_proofs_syscall<E: ExecutorConfig>(
     ctx: &mut SyscallContext<E>,
     _: SyscallCode,
-    word_idx: u32,
-    word: u32,
-) -> Option<u32> {
+    word_idx: u64,
+    word: u64,
+) -> Option<u64> {
     ctx.rt.record.public_values.deferred_proofs_digest[word_idx as usize] = word;
     None
 }
